+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - [V2] Add collection Truncate, Count, Properties, SetProperties
 - [V2] Add and re-organize missing collection properties fields
 - [V1] Expose getters for Context values
+- [V2] Rename CreateCollectionOptions to CreateCollectionProperties
+- [V2] Add support for missing query options (create documents, remove collection, remove view)
 
 ## [1.6.0](https://github.com/arangodb/go-driver/tree/v1.6.0) (2023-05-30)
 - Add ErrArangoDatabaseNotFound and IsExternalStorageError helper to v2